refactured kpi code
