+++ frontend/src/components/kpi/KpiCard.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+type KpiCardProps = {
+    title: string;
+    value: string | number;
+};
+
+export const KpiCard: React.FC<KpiCardProps> = ({ title, value }) => {
+    return (
+        <div className="bg-white rounded-2xl shadow p-5">
+            <p className="text-sm text-slate-500">{title}</p>
+
+            <h2 className="text-2xl font-bold mt-2 text-slate-800">{value}</h2>
+        </div>
+    );
+};
+
+export default KpiCard;
